+++ backend/testing/tester.py
@@ -14,6 +14,7 @@ from kombu import Exchange
 from kombu import Queue
 from kombu import uuid
 from kombu.utils.compat import nested
+from psycopg2 import sql
 
 
 class Tester:
@@ -319,21 +320,38 @@ class Tester:
                 res = db_cur.fetchall()[0]
                 db_con.commit()
                 time.sleep(1)
-            db_cur.close()
-            db_con.close()
 
             Tester.config_request_rpc(connection)
 
             # call all helper functions
-            Helper.resolve(connection, "a", "139.5.46.0/24", "S|0|-", 133720)
-            Helper.mitigate(connection, "b", "139.5.236.0/24")
-            Helper.ignore_hijack(connection, "c", "139.5.237.0/24", "S|0|-", 136334)
-            Helper.comment(connection, "d", "test")
-            Helper.ack_hijack(connection, "e", "true")
-            Helper.multiple_action(connection, ["f", "g"], "mark_seen")
-            Helper.multiple_action(connection, ["f", "g"], "mark_not_seen")
-            Helper.multiple_action(connection, ["h"], "mark_resolved")
-            Helper.multiple_action(connection, ["i"], "mark_ignored")
+            Helper.hijack_resolve(
+                db_con, connection, "a", "139.5.46.0/24", "S|0|-", 133720
+            )
+            Helper.hijack_mitigate(db_con, connection, "b", "10.91.236.0/24")
+            Helper.hijack_ignore(
+                db_con, connection, "c", "139.5.237.0/24", "S|0|-", 136334
+            )
+            Helper.hijack_comment(db_con, connection, "d", "test")
+            Helper.hijack_ack(db_con, connection, "e", "true")
+            Helper.hijack_multiple_action(
+                db_con, connection, ["f", "g"], "hijack_action_acknowledge"
+            )
+            Helper.hijack_multiple_action(
+                db_con, connection, ["f", "g"], "hijack_action_acknowledge_not"
+            )
+            Helper.hijack_multiple_action(
+                db_con, connection, ["h"], "hijack_action_resolve"
+            )
+            Helper.hijack_multiple_action(
+                db_con, connection, ["i"], "hijack_action_ignore"
+            )
+
+            Helper.hijack_delete(
+                db_con, connection, "j", "139.5.237.0/24", "S|0|-", 136334
+            )
+
+            db_cur.close()
+            db_con.close()
 
             for testfile in os.listdir("testfiles/"):
                 self.clear()
@@ -402,7 +420,7 @@ class Tester:
 
 class Helper:
     @staticmethod
-    def resolve(connection, hijack_key, prefix, type_, hijack_as):
+    def hijack_resolve(db_con, connection, hijack_key, prefix, type_, hijack_as):
         hijack_exchange = Exchange(
             "hijack-update", type="direct", durable=False, delivery_mode=1
         )
@@ -415,12 +433,16 @@ class Helper:
                     "hijack_as": hijack_as,
                 },
                 exchange=hijack_exchange,
-                routing_key="resolved",
+                routing_key="resolve",
                 priority=2,
             )
+        result = hijack_action_test_result(db_con, hijack_key, "resolved")
+        assert (
+            result is True
+        ), 'Action "hijack_resolve" for hijack id #{0} failed'.format(hijack_key)
 
     @staticmethod
-    def mitigate(connection, hijack_key, prefix):
+    def hijack_mitigate(db_con, connection, hijack_key, prefix):
         mitigation_exchange = Exchange(
             "mitigation", type="direct", durable=False, delivery_mode=1
         )
@@ -431,9 +453,13 @@ class Helper:
                 routing_key="mitigate",
                 priority=2,
             )
+        result = hijack_action_test_result(db_con, hijack_key, "under_mitigation")
+        assert (
+            result is True
+        ), 'Action "hijack_mitigate" for hijack id #{0} failed'.format(hijack_key)
 
     @staticmethod
-    def ignore_hijack(connection, hijack_key, prefix, type_, hijack_as):
+    def hijack_ignore(db_con, connection, hijack_key, prefix, type_, hijack_as):
         hijack_exchange = Exchange(
             "hijack-update", type="direct", durable=False, delivery_mode=1
         )
@@ -446,12 +472,16 @@ class Helper:
                     "hijack_as": hijack_as,
                 },
                 exchange=hijack_exchange,
-                routing_key="ignored",
+                routing_key="ignore",
                 priority=2,
             )
+        result = hijack_action_test_result(db_con, hijack_key, "ignored")
+        assert (
+            result is True
+        ), 'Action "hijack_ignore" for hijack id #{0} failed'.format(hijack_key)
 
     @staticmethod
-    def comment(connection, hijack_key, comment):
+    def hijack_comment(db_con, connection, hijack_key, comment):
         correlation_id = uuid()
         callback_queue = Queue(
             uuid(),
@@ -477,6 +507,10 @@ class Helper:
             if callback_queue.get():
                 break
             time.sleep(0.1)
+        result = hijack_action_test_result(db_con, hijack_key, "comment", comment)
+        assert (
+            result is True
+        ), 'Action "hijack_comment" for hijack id #{0} failed'.format(hijack_key)
 
     @staticmethod
     def change_conf(connection, new_config, old_config, comment):
@@ -509,35 +543,45 @@ class Helper:
                 time.sleep(0.1)
 
     @staticmethod
-    def ack_hijack(connection, hijack_key, state):
-        correlation_id = uuid()
-        callback_queue = Queue(
-            uuid(),
-            channel=connection.default_channel,
-            durable=False,
-            exclusive=True,
-            auto_delete=True,
-            max_priority=4,
-            consumer_arguments={"x-priority": 4},
+    def hijack_ack(db_con, connection, hijack_key, state):
+        hijack_exchange = Exchange(
+            "hijack-update", type="direct", durable=False, delivery_mode=1
         )
         with connection.Producer() as producer:
             producer.publish(
                 {"key": hijack_key, "state": state},
-                exchange="",
-                routing_key="db-hijack-seen",
-                retry=True,
-                declare=[callback_queue],
-                reply_to=callback_queue.name,
-                correlation_id=correlation_id,
-                priority=4,
+                exchange=hijack_exchange,
+                routing_key="seen",
+                priority=2,
             )
-        while True:
-            if callback_queue.get():
-                break
-            time.sleep(0.1)
+        assert (
+            hijack_action_test_result(db_con, hijack_key, "seen", state) is True
+        ), 'Action "hijack_ack" for hijack id #{0} failed'.format(hijack_key)
 
     @staticmethod
-    def multiple_action(connection, hijack_keys, action):
+    def hijack_delete(db_con, connection, hijack_key, prefix, type_, hijack_as):
+        hijack_exchange = Exchange(
+            "hijack-update", type="direct", durable=False, delivery_mode=1
+        )
+        with connection.Producer() as producer:
+            producer.publish(
+                {
+                    "key": hijack_key,
+                    "prefix": prefix,
+                    "type": type_,
+                    "hijack_as": hijack_as,
+                },
+                exchange=hijack_exchange,
+                routing_key="delete",
+                priority=2,
+            )
+        result = hijack_action_test_result(db_con, hijack_key, "delete")
+        assert (
+            result is True
+        ), 'Action "hijack_delete" for hijack id #{0} failed'.format(hijack_key)
+
+    @staticmethod
+    def hijack_multiple_action(db_con, connection, hijack_keys, action):
         correlation_id = uuid()
         callback_queue = Queue(
             uuid(),
@@ -563,6 +607,42 @@ class Helper:
             if callback_queue.get():
                 break
             time.sleep(0.1)
+
+
+def hijack_action_test_result(db_con, hijack_key, action, extra=None):
+    db_cur = db_con.cursor()
+    query = None
+    query_arguments = None
+    if action == "comment":
+        query = sql.SQL("SELECT COUNT(1) FROM hijacks WHERE key=%s and {}=%s;").format(
+            sql.Identifier(action)
+        )
+        query_arguments = (hijack_key, extra)
+    elif action == "seen":
+        query = sql.SQL("SELECT COUNT(1) FROM hijacks WHERE key=%s and {}=%s;").format(
+            sql.Identifier(action)
+        )
+        query_arguments = (hijack_key, extra)
+    elif action == "delete":
+        query = sql.SQL("SELECT COUNT(1) FROM hijacks WHERE key=%s;")
+        query_arguments = hijack_key
+    else:
+        query = sql.SQL(
+            "SELECT COUNT(1) FROM hijacks WHERE key=%s and {}=true;"
+        ).format(sql.Identifier(action))
+        query_arguments = hijack_key
+    max_tries = 0
+    while max_tries < 20:
+        db_cur.execute(query, query_arguments)
+        res = db_cur.fetchone()
+        db_con.commit()
+        if (res[0] == 1) or (res[0] == 0 and action == "delete"):
+            return True
+
+        time.sleep(1)
+        max_tries += 1
+
+    return False
 
 
 if __name__ == "__main__":